requirements.txt
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.0.4
cloudinary==1.44.0
colorama==0.4.5
dataclasses==0.8
dlib==19.8.1
dnspython==2.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
gunicorn==21.2.0
idna==3.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
opencv-python==4.5.5.64
packaging==21.3
Pillow==8.4.0
pymongo==4.1.1
pyparsing==3.0.7
python-dotenv==0.20.0
requests==2.27.1
six==1.17.0
typing_extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.0.3
zipp==3.6.0
